Reorg functions into files separate files
