+++ src/util/server-util-error.ts
@@ -1,0 +1,8 @@
+export class ServerUtilError extends Error {
+  response: object
+
+  constructor(message: string, response: object) {
+    super(message)
+    this.response = response
+  }
+}
